BIG-19703 Fix Add to cart, Quickview, Foundation, Remove KO
