+++ package.json
@@ -22,9 +22,8 @@
       "browserstate/history.js": "github:browserstate/history.js@^1.8.0",
       "caolan/async": "github:caolan/async@^0.9.2",
       "casperin/nod": "github:casperin/nod@^2.0.4",
-      "foundation": "github:zurb/bower-foundation@^5.5.2",
+      "foundation": "github:bigcommerce-labs/foundation@^5.5.3",
       "jquery": "github:components/jquery@^2.1.3",
-      "knockout": "github:knockout/knockout@^3.3.0",
       "lodash": "npm:lodash@^3.5.0",
       "slick-carousel": "github:kenwheeler/slick@^1.5.5",
       "url": "github:jspm/nodelibs-url@^0.1.0"
